Change feed UI interaction when scrolling to top
Previously, when the user scrolled to the top it would trigger
a reload. I don't think is actually useful in this app since changes
will not come through frequently enough to be important. This fixed a bug in
the case where willDisplayCell trigger is slow enough that it would trigger
two refreshes. This was discovered while fixing a bug with showing the
tableViewFooter that promptly broke the tests. That's the reason for
such a large diff