Disable UI tests
These tests are currently crashing. I think I should just delete these tests,
but I'll spend a bit more time on this. I'm not currently using these tests
to the full extent because it seems difficult to get it all setup with
external libraries.

Reason for the crash: I need to install the relevant libraries in the Podfile
for WiltUITests. Cocoapods-keys doesn't support multiple targets with the
same keys, which makes it unworkable. Relevant issue:
https://github.com/orta/cocoapods-keys/issues/72