Add dismiss functionality to coordinators
This should fix some issues with view updates happening after the view is
disappeared. This may also help with some race conditions around the cache.
Unfortunately now nearly everything needs to be weak, since we don't know
if the view will be deallocated before requests are complete. This makes
sense to me though and the optional handling is simple enough.