Update to iOS 13
This adds support for iOS 13. Mostly I just needed to update the snapshots.

There was one issue with the use of MDCChipView and sizeToFit. From what I can
tell, even though I was calling sizeToFit, it seemed like autolayout was
overriding that value. This seems to be complicated by the reuse of the
cell since it's a UICollectionView. I never actually worked out exactly how
the widths were getting messed up but I ended up activating and deactivating
width constraints based on the label width. I think that sizeToFit isn't
really meant to be used with autolayout anyway