Fix broken test due to incorrect tableview updates
It turned out that this was actually caused by my fake DAO
not reflecting back the delete operation but it made me aware
of a few other changes necessary. Doing batch updates doesn't
seem to mesh well with MVVM at all. I'm sure there's a nicer way to
write this but I'll need to think about it more