Add search feature to listen later
This required adding a new modal using UISearchController
and using a mixture of Firebase and Spotify to do the search.
I need Firebase to retrieve an auth token using my access
token and secret and then I can use the Spotify search freely.

There's a lot of things here that seem like I'm doing this wrong:
- The simulator hangs when running search controller. No idea why. Hence no
snapshot tests
- The activity indicator on the search bar seems hacky and the communication
between the results controller and the search controller seems bad.
UISearchController is meant for filtering an already existing table, so
maybe I should've tried something else
- The animation when it loads seems slow. See above, I probably shouldn't
use UISearchController
- The networking logic is pretty complicated since the user will be
typing a lot. It's probably time to consider adding in something reactive

I've also left a lot of the error handling for now. Sorry, running out
of time to work on this stuff. Will fix this before merging with master.