Add cache for artist activity
I used the FileManager here instead of Core Data as the benefits of
Core Data wouldn't be very useful here. A 10 day cache is used,
I'm tempted to increase this to 1 month since it should only be updated
once a month, however I was worried about user's not seeing the most
recent month until that month was passed.

At some point I'll add some sort of hashing to the cache keys, as I'm
not exactly sure what restrictions there are on artist names in Spotify.
In the worst case, these graphs will just not be cached for now.