Remove Spotify iOS SDK for SpotifyLogin fork
Spotify iOS SDK located here: https://github.com/spotify/ios-sdk
uses PKCE, which means that refresh tokens don't last indefinitely.
To fix this, I've forked SpotifyLogin and am now using that to retrieve
the access token. I've made sure that the client secret is not required,
so that this library can be used in production.

Due to the API change, the flow from view model has changed slightly,
since a controller is needed to start the login popup.

Also, since moving away from Spotify-iOS, it turns out that I was missing
A LOT of imports, so that's most of this commit. Sorry, I should've done
this in two commits.